Docker Action (#2)
Docker Build Action for ARM64 + AMD64

Pushes docker images on pull requests into master, changes to master, and on version tags.

Setup needed is an action variable for the image name in Docker Hub, and action secrets for the Docker Hub username and token.

There's a quirk with the v prefix when tagging in a release:

https://github.com/docker/metadata-action/issues/422